Allow for configuring the location of the hash added to a filename to improve human readability.
The existing functionality adds the hash to the middle of the filename, specifically to just before the first period in the filename.  script.min.js becomes script-a1b2c3...d4e5f6.min.js.  This works without any issues but is a bit ugly for humans to read.  Instead of just hardcoding a change to the hash location, the location is configurable (at the start of the filename, end of the filename with extension copied afterward, or the current "at first period").
 - Start changes script.min.js to a1b2c3...d4e5f6.script.min.js.
 - End changes script.min.js to script.min.js.a1b2c3....d4e5f6.js (the extension is copied to the end to prevent breaking browsers guessing content-type by extension).

Hash location is configured via functional options per [1] and [2]. This method of configuration was chosen because (1) it will not break existing usage of this package (you can still call `hashfs.NewFS()` with a single fs.FS argument), (2) it limits the values for hash location a user can provide to our defined locations, and (3) it allows for future adding of other configurable settings (ex.: hash algorithm).

[1] https://commandcenter.blogspot.com/2014/01/self-referential-functions-and-design.html
[2] https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis